Fortify method for envfit objects

Produces a tidy data frame from the results of an envfit object.

# S3 method for envfit
fortify(model, data, ...)

Arguments

model
an object of class envfit, the result of a call to envfit.

data
additional data to augment the envfit results. Currently ignored.

...
arguments passed to scores.envfit.

Value

A data frame with columns Label, Type, containing the label for, and whether each row refers to, the fitted vector or factor. Remaining variables are coordinates on the respective ordination axes returned by scores.envfit.

Author

Gavin L. Simpson

Examples

data(varespec, varechem)
ord <- metaMDS(varespec)
#> Square root transformation
#> Wisconsin double standardization
#> Run 0 stress 0.1843196 
#> Run 1 stress 0.2331579 
#> Run 2 stress 0.2353272 
#> Run 3 stress 0.2136763 
#> Run 4 stress 0.1955836 
#> Run 5 stress 0.2218502 
#> Run 6 stress 0.2032569 
#> Run 7 stress 0.1825658 
#> ... New best solution
#> ... Procrustes: rmse 0.04163956  max resid 0.1518716 
#> Run 8 stress 0.2269402 
#> Run 9 stress 0.1825658 
#> ... Procrustes: rmse 3.364782e-05  max resid 0.0001012571 
#> ... Similar to previous best
#> Run 10 stress 0.1948413 
#> Run 11 stress 0.2088293 
#> Run 12 stress 0.1969805 
#> Run 13 stress 0.18458 
#> Run 14 stress 0.2404569 
#> Run 15 stress 0.1825658 
#> ... New best solution
#> ... Procrustes: rmse 3.82545e-05  max resid 0.0001221486 
#> ... Similar to previous best
#> Run 16 stress 0.2028828 
#> Run 17 stress 0.2338494 
#> Run 18 stress 0.2396702 
#> Run 19 stress 0.2057245 
#> Run 20 stress 0.2200285 
#> *** Solution reached
fit <- envfit(ord, varechem, perm = 199)

fortify(fit)
#>       Label   Type       NMDS1      NMDS2
#> 1         N Vector -0.02887163 -0.5027736
#> 2         P Vector  0.27285614  0.3455344
#> 3         K Vector  0.32602827  0.2731980
#> 4        Ca Vector  0.43974618  0.4674180
#> 5        Mg Vector  0.41334063  0.5061291
#> 6         S Vector  0.08011005  0.4108088
#> 7        Al Vector -0.63268557  0.3558613
#> 8        Fe Vector -0.62441036  0.2348277
#> 9        Mn Vector  0.57767101 -0.4351840
#> 10       Zn Vector  0.26771528  0.3409638
#> 11       Mo Vector -0.22295679  0.1060218
#> 12 Baresoil Vector  0.46314406 -0.1904459
#> 13 Humdepth Vector  0.67270113 -0.2598610
#> 14       pH Vector -0.31129817  0.3659080

data(dune, dune.env)
ord <- cca(dune)
fit <- envfit(ord ~ Moisture + A1, dune.env, perm = 199)

fortify(fit)
#>       Label     Type        CA1         CA2
#> 1        A1   Vector  0.5560989  0.03376924
#> 2 Moisture1 Centroid -0.7484337 -0.14227818
#> 3 Moisture2 Centroid -0.4651928 -0.21556841
#> 4 Moisture4 Centroid  0.1826781 -0.73151540
#> 5 Moisture5 Centroid  1.1142548  0.57080238