Fortify a "decorana" object.

Fortifies an object of class "decorana" to produce a data frame of the selected axis scores in long format, suitable for plotting with ggplot.

# S3 method for class 'decorana'
fortify(model, data, axes = 1:4, display = c("sites", "species"), ...)

Arguments

model

an object of class "decorana", the result of a call to decorana.

data

currently ignored.

axes

numeric; which axis scores are required?

display

character; the scores to extract in the fortified object.

...

additional arguments passed to scores.decorana.

Value

A data frame in long format containing the ordination scores. The first two components are the axis scores.

Details

TODO

Author

Gavin L. Simpson

Examples

data(dune)

sol <- decorana(dune)
head(fortify(sol))
#>   score label        DCA1       DCA2       DCA3        DCA4
#> 1 sites     1 -0.75835215 -1.3443210 -0.1032421 -0.03611870
#> 2 sites     2 -0.53276038 -0.7711643  0.4688694 -0.29780251
#> 3 sites     3  0.06687385 -0.5015519 -0.0855340 -0.47902825
#> 4 sites     4  0.08025667 -0.2798540  0.6636408 -0.08981166
#> 5 sites     5 -1.03825736 -0.4375810 -0.2073618 -0.24364855
#> 6 sites     6 -0.91509628 -0.1838166 -0.6369058  0.03475827
head(fortify(sol, display = "species"))
#>     score    label       DCA1        DCA2       DCA3       DCA4
#> 1 species Achimill -2.2846487 -2.20135211  1.3366343 -0.7654852
#> 2 species Agrostol  1.6770933  0.34354748  0.8681646 -0.4308604
#> 3 species Airaprae -2.6526545  2.63041715  0.6222752 -1.2834762
#> 4 species Alopgeni  1.0762747 -0.04919152 -0.9947632 -2.0171300
#> 5 species Anthodor -2.5082149  1.36655880  0.8581223 -0.5364167
#> 6 species Bellpere -0.9193113 -0.84068071  1.6094404 -1.0801518