Fortify method for envfit objects

Produces a tidy data frame from the results of an envfit object.

# S3 method for envfit
fortify(model, data, ...)

Arguments

model
an object of class envfit, the result of a call to envfit.

data
additional data to augment the envfit results. Currently ignored.

...
arguments passed to scores.envfit.

Value

A data frame with columns Label, Type, containing the label for, and whether each row refers to, the fitted vector or factor. Remaining variables are coordinates on the respective ordination axes returned by scores.envfit.

Examples

data(varespec, varechem)
ord <- metaMDS(varespec)

#> Square root transformation
#> Wisconsin double standardization
#> Run 0 stress 0.1843196 
#> Run 1 stress 0.2331602 
#> Run 2 stress 0.2353278 
#> Run 3 stress 0.2136773 
#> Run 4 stress 0.1955854 
#> Run 5 stress 0.2218504 
#> Run 6 stress 0.2032569 
#> Run 7 stress 0.1825658 
#> ... New best solution
#> ... Procrustes: rmse 0.04171002  max resid 0.1521984 
#> Run 8 stress 0.2269403 
#> Run 9 stress 0.1825658 
#> ... Procrustes: rmse 0.0001248331  max resid 0.0003366667 
#> ... Similar to previous best
#> Run 10 stress 0.2050757 
#> Run 11 stress 0.2088293 
#> Run 12 stress 0.1969805 
#> Run 13 stress 0.1845801 
#> Run 14 stress 0.2404575 
#> Run 15 stress 0.1825659 
#> ... Procrustes: rmse 0.0001535111  max resid 0.0005017307 
#> ... Similar to previous best
#> Run 16 stress 0.2028828 
#> Run 17 stress 0.2338494 
#> Run 18 stress 0.2396823 
#> Run 19 stress 0.2057245 
#> Run 20 stress 0.2200288 
#> *** Solution reached

fit <- envfit(ord, varechem, perm = 199)

fortify(fit)

#>       Label   Type       NMDS1      NMDS2
#> 1         N Vector -0.02877476 -0.5029069
#> 2         P Vector  0.27272946  0.3455565
#> 3         K Vector  0.32590338  0.2733061
#> 4        Ca Vector  0.43961276  0.4675688
#> 5        Mg Vector  0.41324785  0.5062745
#> 6         S Vector  0.08001655  0.4108906
#> 7        Al Vector -0.63280391  0.3556894
#> 8        Fe Vector -0.62457003  0.2345064
#> 9        Mn Vector  0.57764952 -0.4351441
#> 10       Zn Vector  0.26765710  0.3409615
#> 11       Mo Vector -0.22292058  0.1060425
#> 12 Baresoil Vector  0.46326691 -0.1901603
#> 13 Humdepth Vector  0.67274641 -0.2596255
#> 14       pH Vector -0.31134048  0.3658234

data(dune, dune.env)
ord <- cca(dune)
fit <- envfit(ord ~ Moisture + A1, dune.env, perm = 199)

fortify(fit)

#>       Label     Type        CA1         CA2
#> 1        A1   Vector  0.5560989  0.03376924
#> 2 Moisture1 Centroid -0.7484337 -0.14227818
#> 3 Moisture2 Centroid -0.4651928 -0.21556841
#> 4 Moisture4 Centroid  0.1826781 -0.73151540
#> 5 Moisture5 Centroid  1.1142548  0.57080238