Fortify method for envfit objects

Produces a tidy data frame from the results of an envfit object.

# S3 method for class 'envfit'
fortify(model, data, ...)

Arguments

model

an object of class envfit, the result of a call to envfit.

data

additional data to augment the envfit results. Currently ignored.

...

arguments passed to scores.envfit.

Value

A data frame with columns label, type, containing the label for, and whether each row refers to, the fitted vector or factor. Remaining variables are coordinates on the respective ordination axes returned by scores.envfit.

Author

Gavin L. Simpson

Examples

data(varespec, varechem)
ord <- metaMDS(varespec)
#> Square root transformation
#> Wisconsin double standardization
#> Run 0 stress 0.1843196 
#> Run 1 stress 0.2109009 
#> Run 2 stress 0.1843196 
#> ... Procrustes: rmse 7.555751e-05  max resid 0.0003034326 
#> ... Similar to previous best
#> Run 3 stress 0.2028828 
#> Run 4 stress 0.195049 
#> Run 5 stress 0.2097525 
#> Run 6 stress 0.18584 
#> Run 7 stress 0.1967393 
#> Run 8 stress 0.2353272 
#> Run 9 stress 0.196245 
#> Run 10 stress 0.2298908 
#> Run 11 stress 0.1955836 
#> Run 12 stress 0.197541 
#> Run 13 stress 0.1974418 
#> Run 14 stress 0.2100658 
#> Run 15 stress 0.1948413 
#> Run 16 stress 0.2290939 
#> Run 17 stress 0.239678 
#> Run 18 stress 0.2085949 
#> Run 19 stress 0.2100103 
#> Run 20 stress 0.1974407 
#> *** Best solution repeated 1 times
fit <- envfit(ord, varechem, perm = 199)

fortify(fit)
#>       label   type       NMDS1      NMDS2
#> 1         N Vector -0.02298448 -0.4555495
#> 2         P Vector  0.28785488  0.3044152
#> 3         K Vector  0.33676867  0.2286479
#> 4        Ca Vector  0.39769955  0.3505525
#> 5        Mg Vector  0.41184412  0.4239451
#> 6         S Vector  0.11637297  0.4047729
#> 7        Al Vector -0.60146204  0.3920973
#> 8        Fe Vector -0.54512355  0.3207078
#> 9        Mn Vector  0.58531265 -0.4354945
#> 10       Zn Vector  0.28058970  0.3149280
#> 11       Mo Vector -0.19293587  0.1202130
#> 12 Baresoil Vector  0.43551045 -0.2444282
#> 13 Humdepth Vector  0.69259827 -0.2816765
#> 14       pH Vector -0.40932453  0.3080079

data(dune, dune.env)
ord <- cca(dune)
fit <- envfit(ord ~ Moisture + A1, dune.env, perm = 199)

fortify(fit)
#>       label     type        CA1         CA2
#> 1        A1   Vector -0.5560989 -0.03376924
#> 2 Moisture1 Centroid  0.7484337  0.14227818
#> 3 Moisture2 Centroid  0.4651928  0.21556841
#> 4 Moisture4 Centroid -0.1826781  0.73151540
#> 5 Moisture5 Centroid -1.1142548 -0.57080238