Fortify a "mataMDS" object.

Fortifies an object of class "metaMDS" to produce a data frame of the selected axis scores in long format, suitable for plotting with ggplot.

# S3 method for metaMDS
fortify(model, data, ...)

Arguments

model
an object of class "metaMDS", the result of a call to metaMDS.

data
currently ignored.

...
additional arguments passed to scores.metaMDS. Note you can't use display.

Value

A data frame in long format containing the ordination scores. The first two components are the axis scores.

Details

TODO

Examples

## load example data
data(dune)

ord <- metaMDS(dune)

#> Run 0 stress 0.1192678 
#> Run 1 stress 0.1183186 
#> ... New best solution
#> ... Procrustes: rmse 0.02026992  max resid 0.06495411 
#> Run 2 stress 0.1192679 
#> Run 3 stress 0.1192682 
#> Run 4 stress 0.1183186 
#> ... Procrustes: rmse 3.57951e-05  max resid 9.691533e-05 
#> ... Similar to previous best
#> Run 5 stress 0.1192686 
#> Run 6 stress 0.1192685 
#> Run 7 stress 0.1192679 
#> Run 8 stress 0.1183186 
#> ... Procrustes: rmse 5.277213e-06  max resid 1.492931e-05 
#> ... Similar to previous best
#> Run 9 stress 0.1192678 
#> Run 10 stress 0.1183186 
#> ... Procrustes: rmse 1.081168e-05  max resid 3.468916e-05 
#> ... Similar to previous best
#> Run 11 stress 0.1183186 
#> ... Procrustes: rmse 7.209057e-05  max resid 0.0002266506 
#> ... Similar to previous best
#> Run 12 stress 0.1183186 
#> ... Procrustes: rmse 2.079642e-05  max resid 5.787151e-05 
#> ... Similar to previous best
#> Run 13 stress 0.1192686 
#> Run 14 stress 0.1192685 
#> Run 15 stress 0.119268 
#> Run 16 stress 0.1183186 
#> ... Procrustes: rmse 1.561567e-05  max resid 5.123386e-05 
#> ... Similar to previous best
#> Run 17 stress 0.1183186 
#> ... Procrustes: rmse 3.460559e-05  max resid 9.955366e-05 
#> ... Similar to previous best
#> Run 18 stress 0.1192679 
#> Run 19 stress 0.200356 
#> Run 20 stress 0.1183186 
#> ... Procrustes: rmse 2.214771e-05  max resid 7.007003e-05 
#> ... Similar to previous best
#> *** Solution reached

head(fortify(ord))

#>   Score Label       NMDS1       NMDS2
#> 1 sites     1 -0.84053265 -0.71584691
#> 2 sites     2 -0.50485275 -0.40893645
#> 3 sites     3 -0.08266702 -0.43667555
#> 4 sites     4 -0.11561972 -0.52223972
#> 5 sites     5 -0.62654418 -0.08670082
#> 6 sites     6 -0.54268985  0.11315758