Fortify a "mataMDS" object.

Fortifies an object of class "metaMDS" to produce a data frame of the selected axis scores in long format, suitable for plotting with ggplot.

# S3 method for metaMDS
fortify(model, data, ...)

Arguments

model
an object of class "metaMDS", the result of a call to metaMDS.

data
currently ignored.

...
additional arguments passed to scores.metaMDS. Note you can't use display.

Value

A data frame in long format containing the ordination scores. The first two components are the axis scores.

Details

TODO

Author

Gavin L. Simpson

Examples

## load example data
data(dune)

ord <- metaMDS(dune)
#> Run 0 stress 0.1192678 
#> Run 1 stress 0.1183186 
#> ... New best solution
#> ... Procrustes: rmse 0.02027032  max resid 0.0649614 
#> Run 2 stress 0.1192678 
#> Run 3 stress 0.1192678 
#> Run 4 stress 0.1183186 
#> ... Procrustes: rmse 6.495695e-06  max resid 1.957736e-05 
#> ... Similar to previous best
#> Run 5 stress 0.1192679 
#> Run 6 stress 0.1192679 
#> Run 7 stress 0.1192678 
#> Run 8 stress 0.1183186 
#> ... Procrustes: rmse 4.246485e-06  max resid 1.326091e-05 
#> ... Similar to previous best
#> Run 9 stress 0.1192678 
#> Run 10 stress 0.1183186 
#> ... New best solution
#> ... Procrustes: rmse 1.81361e-06  max resid 4.198697e-06 
#> ... Similar to previous best
#> Run 11 stress 0.1183186 
#> ... Procrustes: rmse 1.450304e-05  max resid 4.526135e-05 
#> ... Similar to previous best
#> Run 12 stress 0.1183186 
#> ... Procrustes: rmse 3.22925e-06  max resid 1.016814e-05 
#> ... Similar to previous best
#> Run 13 stress 0.1192678 
#> Run 14 stress 0.1192679 
#> Run 15 stress 0.1192678 
#> Run 16 stress 0.1183186 
#> ... Procrustes: rmse 1.649744e-06  max resid 3.608095e-06 
#> ... Similar to previous best
#> Run 17 stress 0.1183186 
#> ... Procrustes: rmse 3.850601e-06  max resid 1.198875e-05 
#> ... Similar to previous best
#> Run 18 stress 0.1192678 
#> Run 19 stress 0.2003486 
#> Run 20 stress 0.1183186 
#> ... Procrustes: rmse 6.675886e-06  max resid 2.105997e-05 
#> ... Similar to previous best
#> *** Solution reached
head(fortify(ord))
#>   Score Label       NMDS1       NMDS2
#> 1 sites     1 -0.84052791 -0.71584674
#> 2 sites     2 -0.50485342 -0.40893961
#> 3 sites     3 -0.08266444 -0.43667903
#> 4 sites     4 -0.11561664 -0.52224247
#> 5 sites     5 -0.62654099 -0.08670176
#> 6 sites     6 -0.54269063  0.11315451