Fortify a "mataMDS" object.

Fortifies an object of class "metaMDS" to produce a data frame of the selected axis scores in long format, suitable for plotting with ggplot.

# S3 method for metaMDS
fortify(model, data, ...)

Arguments

model

an object of class "metaMDS", the result of a call to metaMDS.

data

currently ignored.

...

additional arguments passed to scores.metaMDS. Note you can't use display.

Value

A data frame in long format containing the ordination scores. The first two components are the axis scores.

Details

TODO

Author

Gavin L. Simpson

Examples

## load example data
data(dune)

ord <- metaMDS(dune)
#> Run 0 stress 0.1192678 
#> Run 1 stress 0.1192678 
#> ... Procrustes: rmse 3.925717e-05  max resid 0.000119738 
#> ... Similar to previous best
#> Run 2 stress 0.1192678 
#> ... Procrustes: rmse 3.973101e-06  max resid 1.013492e-05 
#> ... Similar to previous best
#> Run 3 stress 0.1183186 
#> ... New best solution
#> ... Procrustes: rmse 0.02027002  max resid 0.06496173 
#> Run 4 stress 0.1192678 
#> Run 5 stress 0.1192679 
#> Run 6 stress 0.1808911 
#> Run 7 stress 0.1183186 
#> ... New best solution
#> ... Procrustes: rmse 1.038131e-05  max resid 3.186343e-05 
#> ... Similar to previous best
#> Run 8 stress 0.1183186 
#> ... New best solution
#> ... Procrustes: rmse 4.581202e-06  max resid 1.379287e-05 
#> ... Similar to previous best
#> Run 9 stress 0.1886532 
#> Run 10 stress 0.1183186 
#> ... New best solution
#> ... Procrustes: rmse 2.32412e-06  max resid 6.667385e-06 
#> ... Similar to previous best
#> Run 11 stress 0.1889653 
#> Run 12 stress 0.1192679 
#> Run 13 stress 0.1183186 
#> ... Procrustes: rmse 5.391472e-06  max resid 1.788081e-05 
#> ... Similar to previous best
#> Run 14 stress 0.1192678 
#> Run 15 stress 0.1183186 
#> ... Procrustes: rmse 3.705225e-06  max resid 1.069455e-05 
#> ... Similar to previous best
#> Run 16 stress 0.1192678 
#> Run 17 stress 0.2192849 
#> Run 18 stress 0.1183186 
#> ... Procrustes: rmse 4.660141e-06  max resid 1.508113e-05 
#> ... Similar to previous best
#> Run 19 stress 0.1192678 
#> Run 20 stress 0.2192919 
#> *** Solution reached
head(fortify(ord))
#>   Score Label       NMDS1       NMDS2
#> 1 sites     1 -0.84053479 -0.71584194
#> 2 sites     2 -0.50485134 -0.40893896
#> 3 sites     3 -0.08266697 -0.43667826
#> 4 sites     4 -0.11562048 -0.52224068
#> 5 sites     5 -0.62654530 -0.08669895
#> 6 sites     6 -0.54269671  0.11315145